4950351007656623852
ManifestID: 4950351007656623852
Date: 2026-06-28T22:57:59+08:00
Version: AT-243285

+++ Assembly-CSharp/DestroyAfter.cs
@@ -23,12 +23,9 @@ public class DestroyAfter : KMonoBehaviour
 
 	private void Update()
 	{
-		if (this.particleSystems != null)
+		if (this.particleSystems != null && !this.IsAlive())
 		{
-			if (!this.IsAlive())
-			{
-				this.DeleteObject();
-			}
+			this.DeleteObject();
 		}
 	}
